+++ parser/src/lex.rs
@@ -612,7 +612,7 @@ pub enum Token {
     MultilineFormatStr(Vec<String>),
     Simple(AsciiToken),
     Glyph(Primitive),
-    Placeholder(usize),
+    Placeholder(Option<usize>),
     Subscr(Subscript),
     Superscr(Superscript),
     LeftArrow,
@@ -678,7 +678,7 @@ impl Token {
             _ => None,
         }
     }
-    pub(crate) fn as_placeholder(&self) -> Option<usize> {
+    pub(crate) fn as_placeholder(&self) -> Option<Option<usize>> {
         match self {
             Token::Placeholder(i) => Some(*i),
             _ => None,
@@ -752,7 +752,8 @@ impl fmt::Display for Token {
             Token::OpenModule => write!(f, "┌─╴"),
             Token::OpenPrivateModule => write!(f, "┌╶╶"),
             Token::CloseModule => write!(f, "└─╴"),
-            Token::Placeholder(i) => write!(f, "^{i}"),
+            Token::Placeholder(Some(i)) => write!(f, "^{i}"),
+            Token::Placeholder(None) => write!(f, "^"),
         }
     }
 }
@@ -774,7 +775,6 @@ pub enum AsciiToken {
     DoubleSemicolon,
     Star,
     Percent,
-    Caret,
     Equal,
     EqualTilde,
     BangEqual,
@@ -802,7 +802,6 @@ impl fmt::Display for AsciiToken {
             AsciiToken::DoubleSemicolon => write!(f, ";;"),
             AsciiToken::Star => write!(f, "*"),
             AsciiToken::Percent => write!(f, "%"),
-            AsciiToken::Caret => write!(f, "^"),
             AsciiToken::Equal => write!(f, "="),
             AsciiToken::BangEqual => write!(f, "!="),
             AsciiToken::EqualTilde => write!(f, "=~"),
@@ -1104,9 +1103,9 @@ impl<'a> Lexer<'a> {
                 "%" => self.end(Percent, start),
                 "^" => {
                     if let Some(x) = self.next_char_if(|c| c.chars().all(|c| c.is_ascii_digit())) {
-                        self.end(Placeholder(x.parse().unwrap()), start)
+                        self.end(Placeholder(Some(x.parse().unwrap())), start)
                     } else {
-                        self.end(Caret, start)
+                        self.end(Placeholder(None), start)
                     }
                 }
                 "=" if self.next_char_exact("~") => self.end(EqualTilde, start),
@@ -1348,9 +1347,15 @@ impl<'a> Lexer<'a> {
                     self.end(Superscr(sup), start)
                 }
                 // Identifiers and unformatted glyphs
-                c if is_custom_glyph(c) || c.chars().all(is_ident_char) || "&!‼".contains(c) => {
+                c if is_custom_glyph(c)
+                    || c.chars().all(is_ident_char)
+                    || "&!‼'′″‴".contains(c) =>
+                {
                     // Get ident start
                     let mut ident = self.ident(start, c).to_string();
+                    while let Some(ch) = self.next_char_if(|ch| "'′″‴".contains(ch)) {
+                        ident.push_str(ch);
+                    }
                     let mut exclam_count = match c {
                         "!" => 1,
                         "‼" => 2,
@@ -1859,7 +1864,7 @@ fn is_formatted_subscript(c: &str) -> bool {
 }
 
 pub(crate) fn canonicalize_ident(ident: &str) -> Ident {
-    canonicalize_subscripts(canonicalize_exclams(ident))
+    canonicalize_subscripts(canonicalize_primes(canonicalize_exclams(ident).as_str()))
 }
 
 /// Rewrite the identifier with the same number of exclamation points
@@ -1880,6 +1885,42 @@ fn place_exclams(ident: &str, count: usize) -> Ident {
     if trailing_single {
         new.push('!');
     }
+    new
+}
+
+/// Rewrite the identifier with the same number of primes
+/// using triple, double, and single prime characters as needed
+fn canonicalize_primes(ident: &str) -> Ident {
+    let mut count = 0;
+    for ch in ident.chars() {
+        count += match ch {
+            '\'' => 1,
+            '′' => 1,
+            '″' => 2,
+            '‴' => 3,
+            _ => 0,
+        };
+    }
+    place_primes(ident, count)
+}
+
+fn place_primes(ident: &str, count: usize) -> Ident {
+    let exclams_removed = ident.trim_end_matches(['!', '‼']);
+    let mut new: Ident = exclams_removed
+        .trim_end_matches(['\'', '′', '″', '‴'])
+        .into();
+    let num_triple = count / 3;
+    let trailing_num = count % 3;
+    for _ in 0..num_triple {
+        new.push('‴');
+    }
+    match trailing_num {
+        0 => {}
+        1 => new.push('′'),
+        2 => new.push('″'),
+        _ => unreachable!(),
+    }
+    new.push_str(&ident[exclams_removed.len()..]);
     new
 }