+++ src/compile/mod.rs
@@ -2780,6 +2780,7 @@ impl Compiler {
         let inputs = self.asm.inputs.clone();
         self.emit_diagnostic_impl(Diagnostic::new(message.into(), span, kind, inputs));
     }
+    #[allow(clippy::print_stdout)]
     fn emit_diagnostic_impl(&mut self, diagnostic: Diagnostic) {
         if self.print_diagnostics {
             eprintln!("{}", diagnostic.report());
@@ -2918,7 +2919,7 @@ impl Compiler {
     /// Bind a function in the current scope
     ///
     /// # Errors
-    /// Returns an error in the binding name is not valid
+    /// Returns an error if the binding name is not valid
     pub fn bind_function(&mut self, name: impl Into<EcoString>, function: Function) -> UiuaResult {
         self.bind_function_with_meta(name, function, BindingMeta::default())
     }
@@ -2942,7 +2943,7 @@ impl Compiler {
     /// This function is the only way to bind `# External!` functions.
     ///
     /// # Errors
-    /// Returns an error in the binding name is not valid
+    /// Returns an error if the binding name is not valid
     pub fn create_bind_function(
         &mut self,
         name: impl Into<EcoString>,